pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codexa"
version = "0.1.0"
description = "Codexa CLI and discovery tooling"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Dynaforge" }]
dependencies = ["PyYAML>=6.0"]

[project.scripts]
codexa = "codexa.cli:main"

[tool.setuptools]
packages = ["codexa", "scripts"]
